requirements.txt
python-dotenv
requests
openai
slack_bolt
python-dotenv
requests_oauthlib
